backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
pydantic>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.17
bcrypt>=4.2.0
apscheduler>=3.10.0
